refactor(93C14D): isolate closeout session binding
