+++ .safeword/hooks/lib/cursor-run-identity.ts
@@ -1,5 +1,4 @@
-import { randomUUID } from 'node:crypto';
-import { existsSync, mkdirSync, readFileSync, renameSync, rmSync, writeFileSync } from 'node:fs';
+import { existsSync, mkdirSync, readFileSync, rmSync, writeFileSync } from 'node:fs';
 import nodePath from 'node:path';
 
 import { hasSafewordProjectMarker, resolveNamespaceRoot } from './namespace-root.js';
@@ -14,7 +13,6 @@ const CODEX_RUN_IDENTITY_CACHE = 'codex-run-identity.json';
 // consumer if they shared a file.
 const CURSOR_REVIEW_STAMP_IDENTITY_CACHE = 'cursor-review-stamp-identity.json';
 const CODEX_REVIEW_STAMP_IDENTITY_CACHE = 'codex-review-stamp-identity.json';
-const CLOSEOUT_BINDING_CACHE = 'closeout-session-binding.json';
 
 /** Fixed cache key for the stamp-helper bridge — not a skill, so not skill-named. */
 const REVIEW_STAMP_CACHE_KEY = 'review-stamp';
@@ -24,16 +22,6 @@ const DEFAULT_MAX_AGE_MS = 5 * 60 * 1000;
 interface ShellRunIdentityCache {
   id: string;
   skillName: string;
-  recordedAt: string;
-}
-
-export interface CloseoutBinding {
-  runtime: 'claude' | 'codex' | 'cursor';
-  id: string;
-  transcriptPath?: string;
-}
-
-interface CloseoutBindingCache extends CloseoutBinding {
   recordedAt: string;
 }
 
@@ -113,15 +101,6 @@ function isReviewStampHelperPath(token: string | undefined): boolean {
   );
 }
 
-function isCloseoutCleanupPath(token: string | undefined): boolean {
-  if (token === undefined) return false;
-  const normalized = token.replaceAll('\\', '/');
-  return (
-    normalized === '.safeword/scripts/closeout-cleanup.ts' ||
-    normalized.endsWith('/.safeword/scripts/closeout-cleanup.ts')
-  );
-}
-
 export function parseRecordSkillInvocationCommand(
   command: string,
 ): { skillName: string } | undefined {
@@ -150,86 +129,6 @@ export function commandInvokesWriteReviewStamp(command: string): boolean {
     const words = commandWords(segment);
     return isBunExecutable(words[0]) && isReviewStampHelperPath(words[1]);
   });
-}
-
-/** True only when an executable shell segment runs the installed closeout guard. */
-export function commandInvokesCloseoutCleanup(command: string): boolean {
-  return splitShellSegments(command).some(segment => {
-    const words = commandWords(segment);
-    return isBunExecutable(words[0]) && isCloseoutCleanupPath(words[1]);
-  });
-}
-
-interface RememberCloseoutBindingInput {
-  projectDirectory: string;
-  runtime: CloseoutBinding['runtime'];
-  id: string | undefined;
-  transcriptPath?: string;
-  now?: Date;
-}
-
-interface ReadFreshCloseoutBindingInput {
-  projectDirectory: string;
-  now?: Date;
-  maxAgeMs?: number;
-}
-
-export function rememberCloseoutBinding(input: RememberCloseoutBindingInput): boolean {
-  const id = nonEmptyString(input.id);
-  if (id === undefined || !hasSafewordProjectMarker(input.projectDirectory)) return false;
-  try {
-    const cachePath = cachePathForProject(input.projectDirectory, CLOSEOUT_BINDING_CACHE);
-    mkdirSync(nodePath.dirname(cachePath), { recursive: true });
-    writeFileSync(
-      cachePath,
-      JSON.stringify({
-        runtime: input.runtime,
-        id,
-        ...(nonEmptyString(input.transcriptPath) === undefined
-          ? {}
-          : { transcriptPath: nonEmptyString(input.transcriptPath) }),
-        recordedAt: (input.now ?? new Date()).toISOString(),
-      } satisfies CloseoutBindingCache),
-      'utf8',
-    );
-    return true;
-  } catch {
-    return false;
-  }
-}
-
-export function readFreshCloseoutBinding(
-  input: ReadFreshCloseoutBindingInput,
-): CloseoutBinding | undefined {
-  const cachePath = cachePathForProject(input.projectDirectory, CLOSEOUT_BINDING_CACHE);
-  if (!existsSync(cachePath)) return undefined;
-  const claimPath = `${cachePath}.claim-${randomUUID()}`;
-  try {
-    // rename(2) is atomic within this directory: concurrent closeout commands
-    // cannot both consume the same short-lived host-session proof.
-    renameSync(cachePath, claimPath);
-    const parsed = JSON.parse(readFileSync(claimPath, 'utf8')) as Partial<CloseoutBindingCache>;
-    const id = nonEmptyString(parsed.id);
-    if (id === undefined || !['claude', 'codex', 'cursor'].includes(parsed.runtime ?? '')) {
-      return undefined;
-    }
-    const recordedAt = Date.parse(parsed.recordedAt ?? '');
-    if (!Number.isFinite(recordedAt)) return undefined;
-    if ((input.now ?? new Date()).getTime() - recordedAt > (input.maxAgeMs ?? DEFAULT_MAX_AGE_MS)) {
-      return undefined;
-    }
-    return {
-      runtime: parsed.runtime as CloseoutBinding['runtime'],
-      id,
-      ...(nonEmptyString(parsed.transcriptPath) === undefined
-        ? {}
-        : { transcriptPath: nonEmptyString(parsed.transcriptPath) }),
-    };
-  } catch {
-    return undefined;
-  } finally {
-    rmSync(claimPath, { force: true });
-  }
 }
 
 /**
